Python/Day 17/requirements.txt
requests
argparse
python-dotenv